entrega_03/requirements.txt
python-dotenv
panel
openai
